Put fade in canvasObject
