--- src/canvasAnimationHelpers.js
@@ -1,27 +1,0 @@
-'use strict';
-
-var canvasAnimationHelpers = {
- fade: function(image, targetOpacity, callback) {
-    var currentOpacity = image.getOpacity();
-    var step = (targetOpacity - currentOpacity) / 30;
-    if (step === 0) {
-      callback();
-      return;
-    }
-
-    var frame = function() {
-      currentOpacity += step;
-      if ((step > 0 && currentOpacity >= targetOpacity) || (step < 0 && currentOpacity <= targetOpacity)) {
-        image.setOpacity(targetOpacity);
-        if (callback) callback();
-        return;
-      }
-
-      image.setOpacity(currentOpacity);
-      OpenSeadragon.requestAnimationFrame(frame);
-    };
-    OpenSeadragon.requestAnimationFrame(frame);
-  }
-};
-
-module.exports = canvasAnimationHelpers;
